scripts/ymls/requirements.txt
pyyaml==6.0.1
sqlparse==0.5.1